ai-service/requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# AI/ML Libraries - Let pip resolve compatible versions
langchain>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=1.0.0
langchain-community
langgraph
google-generativeai

# Google Cloud Services
google-cloud-secret-manager
google-cloud-firestore
google-cloud-storage

# HTTP & Async
httpx>=0.25.0
aiohttp>=3.9.0
aiofiles>=23.2.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Rate Limiting & Caching
slowapi>=0.1.9
python-decouple>=3.8

# Utilities
python-dotenv>=1.0.0

# Monitoring & Logging
structlog>=23.2.0